+++ src/pages/Auth/SignUp.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import { postSignUp } from '../api/requests';
-import useForm from '../hooks/useForm';
-import Form from '../components/auth/Form';
+import useForm from '../../hooks/useForm';
+import Form from '../../components/Auth/Form';
+import { useAuthContext } from '../../context/AuthContext';
 
 export default function SignUp() {
   const navigate = useNavigate();
+  const auth = useAuthContext();
 
   const { values, validValues, handleChange } = useForm({
     initialValues: { email: '', password: '' },
@@ -14,11 +15,12 @@ export default function SignUp() {
   const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
     try {
-      const res = await postSignUp(values);
-      if (res.status === 201) {
-        alert('회원가입이 완료되었습니다. 로그인을 진행해주세요.');
-        navigate('/signin');
-      }
+      const res = await auth.signup(values, () => {
+        if (res.status === 200) {
+          alert('회원가입이 완료되었습니다. 로그인을 진행해주세요.');
+          navigate('/signin');
+        }
+      });
     } catch (err) {
       console.error(err);
     }
