+++ src/pages/Todo/Todo.tsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
-import Input from '../components/todo/Input';
-import { DATA_TEST_ID } from '../constants';
-import { postTodo, getTodoList } from '../api/requests';
-import TodoList from '../components/todo/TodoList';
-import { useTodoContext } from '../context/TodoContext';
+import Input from '../../components/Todo/Input';
+import { DATA_TEST_ID } from '../../constants';
+import { postTodo, getTodoList } from '../../api/requests';
+import TodoList from '../../components/Todo/TodoList';
+import { useTodoContext } from '../../context/TodoContext';
 
 export default function Todo() {
   const { dispatch } = useTodoContext();
